+++ haskell/Probability/Distribution/PhyloCTMC/Properties.hs
@@ -3,7 +3,7 @@ module Probability.Distribution.PhyloCTMC.Properties where
 import Bio.Sequence -- for sequence_to_indices
 import Bio.Alignment
 import Bio.Alphabet  -- for type Alphabet
-import Data.Matrix
+import Numeric.LinearAlgebra
 
 import Data.IntMap (IntMap)
 import qualified Data.IntMap as IntMap