+++ haskell/SModel/Likelihood/FixedA.hs
@@ -5,7 +5,7 @@ import Bio.Alphabet
 import Bio.Alignment
 import Data.BitVector
 import Data.Foldable
-import Data.Matrix
+import Numeric.LinearAlgebra
 import Data.Maybe (maybeToList)
 import Data.Array
 import Foreign.Vector